requirements.txt
Django~=2.2.4
django-bootstrap4==2.3.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-simple-captcha==0.5.12
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
kombu==5.0.2
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.0.0
progress==1.5
prompt-toolkit==3.0.9
psycopg2-binary==2.8.6
pyparsing==2.4.6
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
mysql-connector-python==8.0.23
celery~=5.0.5